Move `joystick.cpp|h` to `src/hardware/input`
