+++ src/gui/mapper.cpp
@@ -21,9 +21,9 @@
 #include "config/config.h"
 #include "config/setup.h"
 #include "gui/mapper.h"
+#include "hardware/input/joystick.h"
 #include "hardware/input/keyboard.h"
 #include "hardware/input/mouse.h"
-#include "hardware/joystick.h"
 #include "hardware/pic.h"
 #include "hardware/timer.h"
 #include "misc/video.h"
