+++ src/gui/sdlmain.cpp
@@ -48,9 +48,9 @@
 #include "gui/sdlmain.h"
 #include "gui/titlebar.h"
 #include "gui_msgs.h"
+#include "hardware/input/joystick.h"
 #include "hardware/input/keyboard.h"
 #include "hardware/input/mouse.h"
-#include "hardware/joystick.h"
 #include "hardware/pic.h"
 #include "hardware/timer.h"
 #include "hardware/video/vga.h"
