+++ src/hardware/meson.build
@@ -18,6 +18,7 @@ libhardware_sources = files(
 
     'input/intel8042.cpp',
     'input/intel8255.cpp',
+    'input/joystick.cpp',
     'input/keyboard.cpp',
     'input/keyboard_scancodes.cpp',
     'input/mouse.cpp',
@@ -69,7 +70,6 @@ libhardware_sources = files(
     'iohandler_containers.cpp',
     'ipx.cpp',
     'ipxserver.cpp',
-    'joystick.cpp',
     'memory.cpp',
     'mpu401.cpp',
     'ne2000.cpp',
