+++ src/ints/bios.cpp
@@ -4,23 +4,23 @@
 
 #include "ints/bios.h"
 
-#include "utils/bitops.h"
-#include "cpu/callback.h"
 #include "config/config.h"
+#include "config/setup.h"
+#include "cpu/callback.h"
 #include "cpu/cpu.h"
+#include "cpu/registers.h"
 #include "dos/dos_memory.h"
 #include "dosbox.h"
 #include "hardware/hardware.h"
-#include "hardware/joystick.h"
-#include "hardware/pic.h"
-#include "hardware/serialport/serialport.h"
-#include "hardware/port.h"
-#include "int10.h"
-#include "utils/math_utils.h"
-#include "hardware/memory.h"
+#include "hardware/input/joystick.h"
 #include "hardware/input/mouse.h"
-#include "cpu/registers.h"
-#include "config/setup.h"
+#include "hardware/memory.h"
+#include "hardware/pic.h"
+#include "hardware/port.h"
+#include "hardware/serialport/serialport.h"
+#include "int10.h"
+#include "utils/bitops.h"
+#include "utils/math_utils.h"
 
 #include <ctime>
 #include <memory>
